requirements.txt
pandas
pytest
pdfplumber
camelot-py[cv]
pytesseract
pdf2image
pikepdf
numpy
opencv-python
pillow
reportlab